+++ cinderlib/exception.py
@@ -21,6 +21,7 @@ VolumeNotFound = exception.VolumeNotFound
 SnapshotNotFound = exception.SnapshotNotFound
 ConnectionNotFound = exception.VolumeAttachmentNotFound
 InvalidVolume = exception.InvalidVolume
+VolumeDriverException = exception.VolumeDriverException
 
 
 class InvalidPersistence(Exception):